refactor: Refactored hotkeys to work around low-level keyboard hook.
